requirements.txt
# Core dependencies
yfinance>=0.2.40
finnhub-python>=2.4.20
twelvedata>=1.2.14
robin-stocks>=3.1.0

# Email
sendgrid>=6.11.0

# Database and caching
redis>=5.0.0
psycopg2-binary>=2.9.9

# Testing
hypothesis>=6.100.0
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
pytz>=2024.1
flask>=3.0.0
beautifulsoup4>=4.12.0
